functions/services-integration/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-identity
azure-storage-blob
azure-eventhub